Remove import *
